Added error field clearing

+++ lib/delayed/worker.rb
@@ -16,6 +16,13 @@ module Delayed
     cattr_accessor :destroy_successfull_jobs
     self.destroy_successfull_jobs = true    
     
+    # only useful if destroy_successful_jobs == false
+    # This will clear out the errors for a successful job
+    # since it succeeded, no reason to keep around
+    cattr_accessor :clear_successful_errors
+    self.clear_successful_errors = false
+      
+    
     self.logger = if defined?(Merb::Logger)
       Merb.logger
     elsif defined?(RAILS_DEFAULT_LOGGER)
@@ -121,7 +128,15 @@ module Delayed
     def run(job)
       runtime =  Benchmark.realtime do
         Timeout.timeout(self.class.max_run_time.to_i) { job.invoke_job }
-        job.destroy
+        
+        if destroy_successful_jobs
+          job.destroy
+        else
+          new_attributes = {:finished_at => Delayed::Job.db_time_now, :failed_at => nil, :locked_at => nil, :locked_by => nil}
+          #sometimes, there is no reason to keep an error message (if the job ended up being successful)
+          new_attributes[:last_error] = nil if clear_successful_errors
+          job.update_attributes(new_attributes)
+        end
       end
       # TODO: warn if runtime > max_run_time ?
       say "* [JOB] #{name} completed after %.4f" % runtime
